requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
spacy==3.7.5
pdfplumber==0.9.0
python-docx==1.1.0
sentence-transformers==2.2.2
openai==0.28.1
numpy==1.24.3
pandas==2.1.3